refactor: BasicBadge 관련 파일을 폴더로 묶습니다.
